move image methods out
